EventFilter: cleanup of event harvester headers, moved text-based harvester to implementation file
